lambda/db_init/requirements.txt
boto3==1.34.80
psycopg2-binary==2.9.9
